requirements.txt
requests
djangorestframework-simplejwt